requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contextlib2==21.6.0
docopt==0.6.2
fastapi==0.88.0
filelock==3.8.2
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
internetarchive==3.0.2
jsonpatch==1.32
jsonpointer==2.3
numpy==1.23.5
packaging==22.0
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
schema==0.7.5
sniffio==1.3.0
starlette==0.22.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
